Update sp.sh: create backup of original file

+++ src/sp.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 cd /Applications/Spotify.app/Contents/Resources/Apps
-cp xpui.spa xpui.spa.bak
+
+if [ ! -f xpui.spa.bak ]; then
+  cp xpui.spa xpui.spa.bak
+fi
+
 unzip -p xpui.spa xpui.js > xpui.js #| sed -e 's/adsEnabled:\!0/adsEnabled:\!1/' -e 's/"allSponsorships"/""/' -e 's/\(.=.=>\)"free"/\1"premium"/' > xpui.js
 
 perl -pi -w -e 's|adsEnabled:!0|adsEnabled:!1|' xpui.js
@@ -8,7 +12,7 @@ perl -pi -w -e 's|allSponsorships||' xpui.js
 # perl -pi -w -e 's/(return|.=.=>)"free"===(.+?)(return|.=.=>)"premium"===/$1"premium"===$2$3"free"===/g' xpui.js
 # perl -pi -w -e 's/\(.=.=>\)"free"/$1"premium"/g' xpui.js
 perl -pi -w -e 's/(case .:|async enable\(.\)\{)(this.enabled=.+?\(.{1,3},"audio"\),|return this.enabled=...+?\(.{1,3},"audio"\))((;case 4:)?this.subscription=this.audioApi).+?this.onAdMessage\)/$1$3.cosmosConnector.increaseStreamTime(-100000000000)/' xpui.js
-perl -pi -w -e 's|.(\?\[.{1,6}[a-zA-Z].leaderboard,)|false$1|' xpui.js
+# perl -pi -w -e 's|.(\?\[.{1,6}[a-zA-Z].leaderboard,)|false$1|' xpui.js
 perl -pi -w -e 's|Enables quicksilver in-app messaging modal",default:\K!.(?=})|false|s' xpui.js
 perl -pi -w -e 's#/a\Kd(?=s/v1)|/a\Kd(?=s/v2/t)|/a\Kd(?=s/v2/se)#b#gs' xpui.js
 perl -pi -w -e 's|(this\._product_state(?:_service)?=(.))|$1,$2.putOverridesValues({pairs:{ads:'\''0'\'',catalogue:'\''premium'\'',product:'\''premium'\'',type:'\''premium'\''}})|' xpui.js
